chore: Refactor Article Native to follow the web.
